+++ Vault/Changes.md
@@ -5,6 +5,7 @@ Supports **Flow v2.4**
 ### New features
 - Added **VaultContext** and **VaultContextWrapper** traits in order to facilitate Vault-based library creation
 - Added **FromIdFactory** and **HasId** traits
+- Added **StorableFactory** trait which facilitates access to database-properties
 
 ## v1.18 - 22.01.2024
 This update focuses on improving grouped (i.e. multi-joined) result parsing.